same_tree/same_tree.go
package same_tree

import "alhoritms/common"

// Given the roots of two binary trees p and q, write a function to check if they are the same or not.
// Two binary trees are considered the same if they are structurally identical, and the nodes have the same value.

// isSameTree проверка на идентичность деревьев
func isSameTree(p *common.TreeNode, q *common.TreeNode) bool {
	if p == nil || q == nil {
		return p == q
	}
	if p.Val != q.Val {
		return false
	}
	return isSameTree(p.Left, q.Left) && isSameTree(p.Right, q.Right)
}
